# apps/web/openapi/notarius.json
{
  "components": {
    "schemas": {
...
      "ArtifactTypeBindingModel": {
        "additionalProperties": false,
        "properties": {
          "artifact_type": {
            "$ref": "#/components/schemas/ArtifactTypeKeyResponse"
          },
          "variable": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Variable",
            "type": "string"
          }
        },
        "required": [
          "variable",
          "artifact_type"
        ],
        "title": "ArtifactTypeBindingModel",
        "type": "object"
      },
...
      "ArtifactTypeKeyResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "schema_version"
        ],
        "title": "ArtifactTypeKeyResponse",
        "type": "object"
      },
...
      "ConfigureNodeSecretRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_graph_revision": {
            "minimum": 1.0,
            "title": "Expected Graph Revision",
            "type": "integer"
          },
          "value": {
            "format": "password",
            "title": "Value",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "value",
          "expected_graph_revision"
        ],
        "title": "ConfigureNodeSecretRequest",
        "type": "object"
      },
...
      "CreateSavedGraphRequest": {
        "additionalProperties": false,
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphEdgeModel"
            },
            "title": "Edges",
            "type": "array"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphNodeModel-Input"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateSavedGraphRequest",
        "type": "object"
      },
...
      "GraphNodeSecretsResponse": {
        "additionalProperties": false,
        "properties": {
          "graph_id": {
            "format": "uuid",
            "title": "Graph Id",
            "type": "string"
          },
          "graph_revision": {
            "title": "Graph Revision",
            "type": "integer"
          },
          "secrets": {
            "items": {
              "$ref": "#/components/schemas/NodeSecretStatusResponse"
            },
            "title": "Secrets",
            "type": "array"
          }
        },
        "required": [
          "graph_id",
          "graph_revision",
          "secrets"
        ],
        "title": "GraphNodeSecretsResponse",
        "type": "object"
      },
...
      "GraphPointModel": {
        "additionalProperties": false,
        "properties": {
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "GraphPointModel",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeSecretStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "name",
          "configured"
        ],
        "title": "NodeSecretStatusResponse",
        "type": "object"
      },
...
      "SavedGraphConversionModel": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version"
        ],
        "title": "SavedGraphConversionModel",
        "type": "object"
      },
...
      "SavedGraphEdgeModel": {
        "additionalProperties": false,
        "properties": {
          "collection_mode": {
            "default": "direct",
            "enum": [
              "direct",
              "map"
            ],
            "title": "Collection Mode",
            "type": "string"
          },
          "conversion_path": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphConversionModel"
            },
            "maxItems": 8,
            "title": "Conversion Path",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "from_node": {
            "maxLength": 255,
            "minLength": 1,
            "title": "From Node",
            "type": "string"
          },
          "from_port": {
            "maxLength": 255,
            "minLength": 1,
            "title": "From Port",
            "type": "string"
          },
          "id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "projection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavedGraphProjectionModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "route_offset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphPointModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_node": {
            "maxLength": 255,
            "minLength": 1,
            "title": "To Node",
            "type": "string"
          },
          "to_plug": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Plug"
          },
          "to_port": {
            "maxLength": 255,
            "minLength": 1,
            "title": "To Port",
            "type": "string"
          }
        },
        "required": [
          "id",
          "from_node",
          "from_port",
          "to_node",
          "to_port"
        ],
        "title": "SavedGraphEdgeModel",
        "type": "object"
      },
...
      "SavedGraphInputPlugModel": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "port": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Port",
            "type": "string"
          }
        },
        "required": [
          "id",
          "port"
        ],
        "title": "SavedGraphInputPlugModel",
        "type": "object"
      },
      "SavedGraphListResponse": {
        "additionalProperties": false,
        "properties": {
          "graphs": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphSummaryResponse"
            },
            "title": "Graphs",
            "type": "array"
          }
        },
        "required": [
          "graphs"
        ],
        "title": "SavedGraphListResponse",
        "type": "object"
      },
...
      "SavedGraphNodeLayoutModel": {
        "additionalProperties": false,
        "properties": {
          "appendix_height": {
            "anyOf": [
              {
                "maximum": 16384.0,
                "minimum": 120.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appendix Height"
          },
          "body_height": {
            "anyOf": [
              {
                "maximum": 16384.0,
                "minimum": 96.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Height"
          },
          "width": {
            "anyOf": [
              {
                "maximum": 16384.0,
                "minimum": 260.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "title": "SavedGraphNodeLayoutModel",
        "type": "object"
      },
      "SavedGraphNodeModel-Input": {
        "additionalProperties": false,
        "properties": {
          "artifact_type_bindings": {
            "items": {
              "$ref": "#/components/schemas/ArtifactTypeBindingModel"
            },
            "title": "Artifact Type Bindings",
            "type": "array"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "input_plugs": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphInputPlugModel"
            },
            "title": "Input Plugs",
            "type": "array"
          },
          "layout": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavedGraphNodeLayoutModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "operator_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Operator Id",
            "type": "string"
          },
          "operator_version": {
            "minimum": 1.0,
            "title": "Operator Version",
            "type": "integer"
          },
          "position": {
            "$ref": "#/components/schemas/GraphPointModel"
          }
        },
        "required": [
          "id",
          "operator_id",
          "operator_version",
          "position"
        ],
        "title": "SavedGraphNodeModel",
        "type": "object"
      },
      "SavedGraphNodeModel-Output": {
        "additionalProperties": false,
        "properties": {
          "artifact_type_bindings": {
            "items": {
              "$ref": "#/components/schemas/ArtifactTypeBindingModel"
            },
            "title": "Artifact Type Bindings",
            "type": "array"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "input_plugs": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphInputPlugModel"
            },
            "title": "Input Plugs",
            "type": "array"
          },
          "layout": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavedGraphNodeLayoutModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "operator_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Operator Id",
            "type": "string"
          },
          "operator_version": {
            "minimum": 1.0,
            "title": "Operator Version",
            "type": "integer"
          },
          "position": {
            "$ref": "#/components/schemas/GraphPointModel"
          }
        },
        "required": [
          "id",
          "operator_id",
          "operator_version",
          "position"
        ],
        "title": "SavedGraphNodeModel",
        "type": "object"
      },
...
      "SavedGraphProjectionModel": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "minItems": 1,
            "title": "Path",
            "type": "array"
          }
        },
        "required": [
          "path"
        ],
        "title": "SavedGraphProjectionModel",
        "type": "object"
      },
      "SavedGraphResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphEdgeModel"
            },
            "title": "Edges",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SavedGraphNodeModel-Output"
            },
            "title": "Nodes",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "SavedGraphResponse",
        "type": "object"
      },
      "SavedGraphSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "edge_count": {
            "title": "Edge Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_count": {
            "title": "Node Count",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "revision",
          "node_count",
          "edge_count",
          "updated_at"
        ],
        "title": "SavedGraphSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Opaque host-only browser session cookie.",
        "in": "cookie",
        "name": "notarius_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Notarius API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{workspace_id}/graphs": {
      "get": {
        "operationId": "list_saved_graphs_v1_workspaces__workspace_id__graphs_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGraphListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Saved Graphs",
        "tags": [
          "saved graphs"
        ]
      },
      "post": {
        "operationId": "create_saved_graph_v1_workspaces__workspace_id__graphs_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSavedGraphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Saved Graph",
        "tags": [
          "saved graphs"
        ]
      }
    },
...
    "/v1/workspaces/{workspace_id}/graphs/{graph_id}/node-secrets": {
      "get": {
        "operationId": "get_node_secret_status_v1_workspaces__workspace_id__graphs__graph_id__node_secrets_get",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Graph Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphNodeSecretsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Node Secret Status",
        "tags": [
          "node secrets"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/graphs/{graph_id}/nodes/{node_id}/secrets/{name}": {
      "delete": {
        "operationId": "delete_node_secret_v1_workspaces__workspace_id__graphs__graph_id__nodes__node_id__secrets__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Graph Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_graph_revision",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Graph Revision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Node Secret",
        "tags": [
          "node secrets"
        ]
      },
      "put": {
        "operationId": "configure_node_secret_v1_workspaces__workspace_id__graphs__graph_id__nodes__node_id__secrets__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "graph_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Graph Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureNodeSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeSecretStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Configure Node Secret",
        "tags": [
          "node secrets"
        ]
      }
    },
...
  }
}